New fields to `PersonContact`
Add first_name and last_name fields to `PersonContact` since
it makes sense.

It is justified to have both first name and last name required
because the persons with one name are minority in the world.  See
http://en.wikipedia.org/wiki/Mononymous_person.

Also make `create_random_person` to set the first name and last name
also and update the tests.

Refs SHOOP-832